Move debug_builtins
